requirements.txt
pandas
numpy
yfinance
alpaca-py
schwab-py
backtrader
pandas-ta
pytest
scikit-learn
python-dotenv
ruff
black
mypy
streamlit
plotly
